major refactoring
broke into a main package and a package which could be used for example by a web server backend to run the program using an HTML form and JS on the frontend could display the output in a pretty way. To that end, I added the ability to return the summary JSON encoded. I also added some user defined types and played with some cool golang features.